requirements.txt
asgiref==3.8.0
Django==4.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
Markdown==3.6
mysqlclient==2.1.1
python-dotenv==1.0.0
sqlparse==0.4.4